Repository Pattern

+++ src/main/java/com/mani/example/productservices/Service/ProductService.java
@@ -9,7 +9,7 @@ import com.mani.example.productservices.Model.Product;
 import java.util.List;
 
 public interface ProductService {
-    public Product getProductById(Long id);
+    public Product getProductById(Integer id);
 
     public Product createProduct(String title,
                               String description,
